# Conflicts:
#	sql/base/auth_database.sql
#	sql/base/characters_database.sql
#	src/server/database/Database/Implementation/CharacterDatabase.cpp
#	src/server/game/DataStores/DBCEnums.h
#	src/server/game/Entities/Item/Item.cpp
#	src/server/game/Entities/Player/Player.cpp
#	src/server/game/Miscellaneous/SharedDefines.h
#	src/server/game/Server/Protocol/Opcodes.h
#	src/server/game/Server/WorldSession.cpp
#	src/server/game/Server/WorldSession.h

+++ src/server/game/Guilds/Guild.h
@@ -251,7 +251,7 @@ struct GuildReward
 {
     uint32 ItemID;
     uint8 MinGuildRep;
-    int32 RaceMask;
+    uint64 RaceMask;
     uint64 Cost;
     std::vector<uint32> AchievementsRequired;
 };
